srcsetfill.html


<!doctype html>
<!--[if lte IE 8]> <html lang="en" class="ie-lte8"> <![endif]-->
<!--[if (gt IE 9)|!(IE)]><!--> <html lang="en"> <!--<![endif]-->
<head>
	<meta charset="utf-8">
	<title>ResponsiveImages.org</title>
	<meta name="description" content="Our goal is a markup-based means of delivering alternate image sources based on device capabilities.">
	<meta property="og:url" content="http://responsiveimages.org">
	<meta property="og:site_name" content="Responsive Images Community Group">
	<meta property="og:type" content="website">
	<meta property="og:title" content="Responsive Images Community Group">
	<meta property="og:image" content="http://responsiveimages.org/img/ricg-icon.png">
	<meta property="og:description" content="Read the latest news on responsive images, and get involved!">


	<meta name="viewport" content="width=device-width,initial-scale=1">

	<link rel="stylesheet" href="css/enhanced.css" media="only all">
</head>
<body>

	<header class="doc-head">
			<h1 class="logo"><img src="img/logo.png" alt="Responsive Images Community Group logo"></h1>
			<p class="intro">Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
			tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
			quis nostrud exercitation ullamco laboris nisi ut aliquip.</p>
	</header>

	<div class="main srcsetfill">
		<h1 class="hed">An <code>srcset</code> parser</h1>
		<p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
		tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
		quis nostrud exercitation ullamco laboris nisi.</p>	
		<form id="srcsetTester">
			<fieldset>
				<legend class="demo-hed">Element</legend>
				<code>&lt;img|picture|source </code>
				<label><code>src=</code>
					<input id="src" value="default.png" placeholder="default.png">
				</label>
				<label>
					<code>srcset=</code>
					<input id="srcset" placeholder="foo 2x, bar 200w, bas 400h">
			</label><code>&gt;</code>
			</fieldset>
			<fieldset>
				<legend class="demo-hed">Viewport (read only!)</legend>
				<label>Width:  <input id="width" min="1" type="number" readonly> </label>
				<label>Height: <input id="height" min="1" type="number" readonly></label>
				<label>Density:<input id="density" min="0" step=".1" type="number" readonly></label>
			</fieldset>
			<fieldset id="output">
				<legend class="demo-hed">Output</legend>
				<label>Best match: 
					<output id="out"></output>
				</label>
			</fieldset>
		</form>

<script>
(function (window) {
    "use strict";
    window.addEventListener("DOMContentLoaded", setUpTester);

    function setUpTester() {
        var srcsetTester = document.getElementById("srcsetTester"),
            elem = document.createElement("x-element"),
            attrValues = [{
                name: "srcset",
                value: ""
            }, {
                name: "src",
                value: "default.png"
            }],
            srcsetAttr;

        //create fake IDL attributes for x-element
        attrValues.forEach((function () {
            return function (prop) {
                setupAttributes(prop);
                //initialize attribute
                elem[prop.name] = prop.value;
            }
        }()));

        srcsetAttr = elem.getAttributeNode("srcset");
 		
 		//Wire up events
 		window.addEventListener("resize", popualteForm)
        srcsetTester.addEventListener("input", findSrcset);
        
        //run
        popualteForm();
 
        function setupAttributes(prop) {
            var props = {
                get: function () {
                    return prop.value;
                },
                set: function (value) {
                    elem.setAttribute(prop.name, String(value));
                }
            }
            Object.defineProperty(elem, prop.name, props);
            srcsetTester[prop.name].addEventListener("keyup", function (e) {
                elem[prop.name] = e.target.value;
            });
        }

        //populate form
        function popualteForm() {
            srcsetTester.width.value = window.innerWidth;
            srcsetTester.height.value = window.innerHeight;
            srcsetTester.density.value = window.devicePixelRatio | 1.0;
            findSrcset();
        }

        function findSrcset() {
            var result = window.srcsetParser.parse(srcsetAttr)
            showResult(result);
        }

        function showResult(value) {
            srcsetTester.out.value = JSON.stringify(value);
        }
    }
}(window));
</script>
<div id="log"></div>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="srcsetfill.js"></script>
<script>
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', '');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, null);
        assert_equals(result.density, undefined);
    }, 'If candidates is empty, return null and undefined.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', '\t \t\r\f\f\r \n \t');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, null);
        assert_equals(result.density, undefined);
    }, 'Ignore whitespace at the attribute value.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'pass 1x, fail 1x');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Ignore duplicates, even if the url is different.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'pass 0002.000x, b 02x, b 02.0000x, fail 2x');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 2);
    }, 'Ignore a long list of duplicates with different densities.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'pass 0002.000h, a 02h, c 02.0000h, fail 2h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Ignore a long list of duplicates with the same heights.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'pass 0002.000h 01w 1x, a 02h 001w, c 02.0000h, fail 2h 001w 1x');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Ignore a long list of duplicates with the same width and height and density.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'fail 0002h 01w 1x, a 02h 001w 1x, c 02h 0001w 1x, fail 2h 001w 1x, pass 2000w');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Ignore a long list of duplicates with the same width and height and density.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('srcset', 'fail 0002h 01w 1x, a 02h 001w 1x, c 02h 0001w 1x, fail 2h 001w 1x, pass 2000w');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Ignore a long list of duplicates with the same width and height and density.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'pass');
        img.setAttribute('srcset', 'fail 1w, fail 2w, fail 3w');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test that src attribute is selected as a candidate.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'fail1 1w, pass 2000w, fail2 2000w, fail3 3w');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test that src attribute is not selected as a candidate.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'pass 2000w, fail1 3000w, fail2 4000w');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test selection of smallest large width.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'pass 2000h, fail1 3000h, fail2 4000h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test selection of smallest large height.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'fail1 3000h, fail2 4000h, pass 2000h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test selection of smallest large height.');
    test(function () {
        var img = new Image(''),
            result;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'fail1 3000h, fail2 4000h, pass 2000h, fail1 3000h, fail 2000h, fail2 4000h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test selection of smallest large height, with redudant values.');
    test(function () {
        var img = new Image(''),
            result,
            width = window.innerWidth + 'w';
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'pass ' + width, 'fail ' + width);
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, "Test selection of the right source when window\'s width is the same as the source.");
    test(function () {
        var img = new Image(''),
            result,
            height = window.innerHeight;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'pass ' + height + 'h', 'fail ' + (height - 1) + 'h', 'fail ' + (height + 1) + 'h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, "Test selection of the right source when window\'s width is the same as the source and other bigger and smaller sources are available.");
    test(function () {
        var img = new Image(''),
            result,
            height = window.innerHeight;
        img.setAttribute('src', 'fail_src');
        img.setAttribute('srcset', 'pass ' + height + 'h', 'fail ' + (height - 1) + 'h', 'fail ' + (height + 1) + 'h');
        result = window.srcsetParser.parse(img.getAttributeNode('srcset'));
        assert_equals(result.url, 'pass');
        assert_equals(result.density, 1);
    }, 'Test selection of the right source when window\'s height is the same as the source and other bigger and smaller sources are available.');
</script>
</div>
<div class="doc-foot">

			<aside class="assets">
				<a href="RICG-Logos.zip">
					<img src="img/ricg-icon.png" class="assets-thumb">
					<h1 class="assets-hed">Download the RICG Logos</h1>
					<p>Want to show your support for the goals of the RICG? Use our logo!</p>
				</a>
			</aside>

			<nav class="social">
				<h1 class="a11y">Share this page</h1>
				<a class="subhed twitter" href="http://twitter.com/home?status=Read%20the%20latest%20news%20on%20responsive%20images%2C%20and%20get%20involved!%0Ahttp%3A%2F%2Fresponsiveimages.org">Twitter</a>
				<a class="subhed facebook" href="https://www.facebook.com/sharer.php?u=http%3A%2F%2Fresponsiveimages.org&amp;t=Read%20the%20latest%20news%20on%20responsive%20images%2C%20and%20get%20involved!">Facebook</a>
				<a class="subhed g-plus" href="https://plus.google.com/share?url=http%3A%2F%2Fresponsiveimages.org">Google Plus</a>
			</nav>

		</div>
	<script>var _gaq=[['_setAccount','UA-35760540-1'],['_trackPageview'],['_setDomainName', 'responsiveimages.org']];(function(d,t){var g=d.createElement(t),s=d.getElementsByTagName(t)[0];g.src='//www.google-analytics.com/ga.js';s.parentNode.insertBefore(g,s)}(document,'script'))</script>
